voice-agentwithopenaisdk/requirements.txt
streamlit>=1.33.0
openai>=1.22.0
python-dotenv>=1.0.1
qdrant-client>=1.8.0
langchain>=0.1.13
langchain-community>=0.0.27
pypdf>=3.17.1
fastembed>=0.1.0
asyncio
uuid
fpdf>=1.7.2
ffmpeg-python>=0.2.0
torch>=2.1.0
whisper @ git+https://github.com/openai/whisper.git
